@solnic 2015-01-15
Thanks, I'm gonna update and cleanup devel/test setup so that build is passing in master, then could you merge master into your branch and then I could merge it, OK?