@solnic 2015-01-16
BTW you could also check out another PR (that I must've missed for some reason :() that's related to performance