@tb0yd 2015-01-22
I couldn't reproduce his findings